+++ envtestutils/apiserver/apiserver.go
@@ -31,7 +31,7 @@ import (
 	"time"
 
 	"github.com/onmetal/controller-utils/buildutils"
-	"github.com/onmetal/controller-utils/envtestutils/testing/controlplane"
+	"github.com/onmetal/controller-utils/internal/testing/controlplane"
 	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/rest"
